% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\docType{data}
\name{brownianbridge.empcdf}
\alias{brownianbridge.empcdf}
\title{brownianbridge.empcdf}
\format{An object of class \code{NULL} of length 0.}
\usage{
brownianbridge.empcdf
}
\description{
A global reference to empcdf.ref.
It is used an empirical quantile function to determine 
p-values in the asymptotic wasserstein test function wasserstein.test.asy
}
\keyword{data}
