% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wasserstein_SingleCell.r
\name{CombinePVal}
\alias{CombinePVal}
\title{CombinePVal}
\usage{
CombinePVal(r, s)
}
\arguments{
\item{r}{vector of length N of the p-values corresponding to the first test}

\item{s}{vector of length N of the p-values corresponding to the second test}
}
\value{
A vector of length N of the combined p-values
}
\description{
For a given set of N pairs of p-values, aggregates each respective pair of 
p-values into a combined p-value according to Fisher’s method
}
\details{
For a given set of pairs of p-values, aggregates each respective 
pair of p-values into a combined p-value according to Fisher’s method.
Applies the fishersCombinedPVal function to a whole set of N pairs of
p-values.
}
\examples{
r<-c(0.007,6e-23,0.77,0.23,2e-22)
s<-c(1e-6,3e-54,0.55,0.001,4e-5)
CombinePVal(r,s)


}
