% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{concat}
\alias{concat}
\title{Concatenate Numeric Vectors}
\usage{
concat(x_, y_)
}
\arguments{
\item{x}{NumericVector first vector to be written to the output}

\item{y}{NumericVector second vector to be written to the output}
}
\value{
concatenation of y after x
}
\description{
`concat` returns a NumericVector that represents the concatenation of two input vectors.
The elements of the second given vector are written to the output vector after the elements
in the first vector
}
