% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cumSum}
\alias{cumSum}
\title{Cumulative Sum}
\usage{
cumSum(x_, last_index = 0L)
}
\arguments{
\item{x}{NumericalVector}
}
\value{
a vector containing cumulative sums of the values
	in the input vector
}
\description{
Returns the cumulative sums of a NumericalVector object.
}
