% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PValues.R
\name{.fishersCombinedPval}
\alias{.fishersCombinedPval}
\title{Compute combined p-value using Fisher's method}
\usage{
.fishersCombinedPval(x)
}
\arguments{
\item{x}{vector of the two p-values that are to be aggregated}
}
\value{
The combined p-value
}
\description{
Aggregates two p-values into a combined p-value according to Fisher’s method
}
\details{
Aggregates two p-values into a combined p-value according to
Fisher’s method
}
