% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PValues.R
\name{.gdpFittedPValue}
\alias{.gdpFittedPValue}
\title{.gdpFittedPValue}
\usage{
.gdpFittedPValue(val, distr.ordered, pvalue)
}
\arguments{
\item{val}{atomic result of an observation}

\item{distr.ordered}{distribution of random atomic observations}

\item{pvalue}{estimated p value of this observation}
}
\value{
pvalue.adjusted, Anderson-Darling p-value of gdp fitting, number
 of exceedances for calculating a pvalue from the empirical distribution
}
\description{
Fitting Procedure to correct P a pvalue
}
