% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{.quantileCorrelation}
\alias{.quantileCorrelation}
\title{Compute the quantile-quantile correlation}
\usage{
.quantileCorrelation(x, y, pr = NULL)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{pr}{levels at which the quantiles of \eqn{x} and \eqn{y} are computed; by default, 1000 equidistant quantiles at levels \eqn{\frac{k-0.5}{1000}}, where \eqn{k=1,\ldots,1000}, are used}
}
\value{
quantile-quantile correlation of \eqn{x} and \eqn{y}
}
\description{
Computes the quantile-quantile correlation of two samples \eqn{x} and \eqn{y}, using the quantile
type 1 implementation in R and the Pearson correlation.
}
