% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{.relativeError}
\alias{.relativeError}
\title{.relativeError}
\usage{
.relativeError(x, y)
}
\arguments{
\item{x}{numerical (in the sense of is.numeric)}

\item{y}{numerical (in the sense of is.numeric)}
}
\value{
The relative error between x and y
}
\description{
Computes the relative error between two numericals (in the sense of
is.numeric) x and y.
If x and y are vectors, it is assumed that length(x) == length(y).
}
\details{
The relative error e is defined as: e = | 1 -  x/y |
}
