% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WassersteinSingleCell.R
\name{.testWass}
\alias{.testWass}
\title{.testWass}
\usage{
.testWass(dat, condition, permnum, inclZero = TRUE)
}
\arguments{
\item{dat}{matrix of single-cell RNA-sequencing expression data with genes in
rows and samples (cells) in columns}

\item{condition}{vector of condition labels}

\item{permnum}{number of permutations used in the permutation testing
procedure}

\item{inclZero}{logical; if TRUE, the one-stage method (i.e. semi-parametric
testing applied to all (zero and non-zero) expression values) is performed;
if FALSE, the two-stage method (i.e. semi-parametric testing applied to
non-zero expression values only, combined with a separate testing for
differential proportions of zero expression using logistic regression) is
performed. Default is TRUE}
}
\value{
matrix with every row being the wasserstein test of one gene between
the two conditions.
 See the corresponding values in the description of the function
wasserstein.sc, where the argument inclZero=TRUE in .testWass has to be
identified with the argument method=”OS”, and the argument inclZero=FALSE in
.testWass with the argument method=”TS”.
}
\description{
Two-sample test for single-cell RNA-sequencing data to check for differences
between two distributions (conditions) using the 2-Wasserstein distance:
Semi-parametric implementation using a permutation test with a generalized 
Pareto distribution (GPD) approximation to estimate small p-values accurately
}
\details{
Details concerning the permutation testing procedures for
single-cell RNA-sequencing data can be found in Schefzik and 
Goncalves (2019).
}
\references{
Schefzik and Goncalves 2019
}
