% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\docType{data}
\name{empcdf.ref}
\alias{empcdf.ref}
\title{An empirical cumulative distribution function based on the values in
value.integral}
\format{An object of class \code{NULL} of length 0.}
\usage{
empcdf.ref
}
\description{
An empirical cumulative distribution function based on the values in
value.integral
}
\keyword{data}
