% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\docType{data}
\name{global.empcdf.ref}
\alias{global.empcdf.ref}
\title{A global reference to empcdf.ref}
\format{An object of class \code{NULL} of length 0.}
\usage{
global.empcdf.ref
}
\description{
A global reference to empcdf.ref
}
\keyword{data}
