% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{interval_table}
\alias{interval_table}
\title{Interval Table}
\usage{
interval_table(datavec, interval_breaks, init_value = 0L)
}
\arguments{
\item{datavec}{NumericVector with elements to be distributed over the intervals}

\item{interval_breaks}{NumericVector with n interval_borders that are
interpreted as interval breaks:\cr 
(-Inf, breaks[0]], (breaks[0], breaks[1]), ... , (breaks(n), Inf)}

\item{ini_value}{Default frequency that is assigned to each interval_breaks.
Counted interval frequencies are added to the default frequency.}
}
\value{
The frequency with which elements of datavec fall into each of the intervals defined
		by the second argument interval_breaks
}
\description{
Given a NumericVector datavec and a NumericVector of interval breaks,
a table with the number of elements in datavec that fall into each of the
intervals is returned.
}
