% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rep_weighted}
\alias{rep_weighted}
\title{Repeat weighted}
\usage{
rep_weighted(x, freq_table)
}
\arguments{
\item{x}{NumericVector object with the elements that are to be weighted}

\item{freq_table}{NumericVector object representing a weight vector.
Each element in x is repeated accoring to its weight in freq_table}
}
\value{
the weight-repeated NumericVector of x
}
\description{
Returns a weighted NumericVector of a given input vector.
Each element x[i] in the given input vector x is repeated according
to the weight vector at position i
}
