% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sq_wasserstein}
\alias{sq_wasserstein}
\title{sq_wasserstein}
\usage{
sq_wasserstein(a_, b_, p = 1)
}
\arguments{
\item{p}{exponent of the wasserstine distance}

\item{a}{Vector representing an empirical distribution under condition A}

\item{b}{Vector representing an empirical distribution under condition B}
}
\value{
The squared wasserstein distance between a and b
}
\description{
Squared Wasserstein distance between two vectors.
}
\details{
The squared wasserstein distance has useful properties for decomposition.
}
