% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sq_wasserstein_decomp}
\alias{sq_wasserstein_decomp}
\title{sq_wasserstein_decomp
Squared Wasserstein distance between two vectors, decomposed into size, location and shape.}
\usage{
sq_wasserstein_decomp(a_, b_, p = 1)
}
\arguments{
\item{p}{exponent of the wasserstine distance}

\item{a}{Vector representing an empirical distribution under condition A}

\item{b}{Vector representing an empirical distribution under condition B}
}
\value{
The wasserstein distance between a and b, decomposed into terms for size, location, and shape
}
\description{
sq_wasserstein_decomp
Squared Wasserstein distance between two vectors, decomposed into size, location and shape.
}
