% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WassersteinSingleCell.R
\docType{methods}
\name{testZeroes}
\alias{testZeroes}
\title{testZeroes}
\usage{
testZeroes(x, y, ...)
}
\arguments{
\item{x}{matrix of single-cell RNA-sequencing expression data with genes in
rows and samples (cells) in columns}

\item{y}{vector of condition labels}

\item{...}{further arguments that can include:
\itemize{
\item *these* vector of row numbers (i.e. gene numbers) employed to test for
 differential proportions of zero expression. Default is seq_len(dat)
}}
}
\value{
A vector of (unadjusted) p-values
}
\description{
Test for differential proportions of zero expression between two conditions
for a specified set of genes
}
\details{
Test for differential proportions of zero expression between two
conditions that is not explained by the detection rate using a (Bayesian)
logistic regression model. Adapted from the scDD package (Korthauer et al.
2016).
}
\examples{
x1 <- c(rnorm(100,42,1), rnorm(102,45,3))
x2 <- c(rnorm(100,0,1), rnorm(102,0,2))
dat <- matrix(c(x1,x2), nrow=2, byrow=True)
condition <- c(rep(1,100), rep(2,102))

# test over all rows
testZeroes(dat, condition)
# only consider the second row
testZeroes(dat, condition, these=c(2))

}
\references{
Korthauer et al. (2016).
}
