% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wasserstein_SingleCell.r
\name{testZeroes}
\alias{testZeroes}
\title{testZeroes}
\usage{
testZeroes(dat, cond, these = 1:nrow(dat))
}
\arguments{
\item{dat}{matrix of single-cell RNA-sequencing expression data with genes in rows and samples (cells) in columns}

\item{these}{vector of row numbers (i.e. gene numbers) employed to test for differential proportions of zero expression; default is 1:nrow(dat)}

\item{condition}{vector of condition labels}
}
\value{
A vector of (unadjusted) p-values
}
\description{
Test for differential proportions of zero expression between two conditions for a specified set of genes
}
\details{
Test for differential proportions of zero expression between two conditions that is not explained by the detection rate using a (Bayesian) logistic regression model. Adapted from the scDD package (Korthauer et al. 2016).
}
\examples{
dat<-c(100, 120, 100, 90, 0, 0, 90, 80, 10, 50)
condition<-c(1, 2, 1, 2, 1, 2, 1, 2, 1, 2)
testZeroes(dat,condition)


}
\references{
Korthauer et al. (2016).
}
