% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wasserstein}
\alias{wasserstein}
\title{wasserstein
Wasserstein distance between two vectors.}
\usage{
wasserstein(a_, b_, p = 1)
}
\arguments{
\item{p}{exponent of the wasserstine distance}

\item{a}{Vector representing an empirical distribution under condition A}

\item{b}{Vector representing an empirical distribution under condition B}
}
\value{
The wasserstein distance between a and b
}
\description{
wasserstein
Wasserstein distance between two vectors.
}
