% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasserstein_single_cell.r
\name{wasserstein.sc}
\alias{wasserstein.sc}
\title{Testing procedure for single-cell data based on the wasserstein distance}
\usage{
wasserstein.sc(dat, condition, seedex, permnum, method)
}
\arguments{
\item{dat}{single cell expression values for two conditions}

\item{condition}{vector of the same length as dat, describing which \cr
condition every datapoint is associated to}

\item{seedex}{seed for generating reproducable distributions}

\item{permnum}{number of permutations that should be produced in calculating the p-values}

\item{method}{one of "OS" or "TS", describing wether a one-stage ("OS") or
two-stage ("TS") test should be performed. \cr
TS - the distribution of the zero values is
independently tested for differential proportions of zeroes \cr
OS - all datapoints are tested with the wasserstein distance derived model}
}
\value{
Test statistics
}
\description{
wasserstein.sc wraps around the testWass function to provide a interface for calling
wasserstein distance based test of differential expression in sigle cell RNA data ...
}
\examples{
dat <- c(100, 120, 100, 90, 0, 0, 90, 80, 10, 50))
condition <- c(1, 2, 1, 2, 1, 2, 1, 2, 1, 2)
wasserstein.sc(dat, condition, 24, 10000, "OS")


}
