% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasserstein_test.r
\name{wasserstein.test.asy}
\alias{wasserstein.test.asy}
\title{wasserstein.test.asy}
\usage{
wasserstein.test.asy(x, y)
}
\arguments{
\item{x}{univariate sample (vector) representing the distribution of
condition A}

\item{y}{univariate sample (vector) representing the distribution of
condition B}
}
\value{
A vector concerning the testing results, precisely (see Schefzik and
Goncalves (2019) for details)
\itemize{
\item d.wass: 2-Wasserstein distance between the two samples computed
by quantile approximation
\item d.wass^2 squared 2-Wasserstein distance between the two samples
computed by quantile approximation
\item d.comp^2: squared 2-Wasserstein distance between the two samples
computed by decomposition approximation
\item d.comp: 2-Wasserstein distance between the two samples computed by
decomposition approximation
\item location: location term in the decomposition of the squared
2-Wasserstein distance between the two samples
\item size: size term in the decomposition of the squared 2-Wasserstein
distance between the two samples
\item shape: shape term in the decomposition of the squared 2-Wasserstein
distance between the two samples
\item rho: correlation coefficient in the quantile-quantile plot
\item pval: p-value of the 2-Wasserstein distance-based test using
asymptotic theory
\item perc.loc: fraction (in %) of the location part with respect to the
overall squared 2-Wasserstein distance obtained by the decomposition
approximation
\item perc.size: fraction (in %) of the size part with respect to the
overall squared 2-Wasserstein distance obtained by the decomposition
approximation
\item perc.shape: fraction (in %) of the shape part with respect to the
overall squared 2-Wasserstein distance obtained by the decomposition
approximation
\item decomp.error: relative error between the squared 2-Wasserstein
distance computed by the quantile approximation and the squared
2-Wasserstein distance computed by the decomposition approximation
}
}
\description{
Two-sample test to check for differences between two distributions
(conditions) using the 2-Wasserstein distance: Implementation using a test
based on asymptotic theory
}
\details{
This is the asymptotic version of wasserstein.test, for the
semi-parametric procedure see wasserstein.test.sp

Details concerning the testing procedure based on asymptotic theory
can be found in Schefzik and Goncalves (2019).
}
\examples{
set.seed(32)
x<-rnorm(500)
y<-rnorm(500,2,1.5)
wasserstein.test.asy(x,y)


}
\references{
Schefzik, R. and Goncalves, A. (2019).
}
