% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\docType{data}
\name{brownianBridgeEmpcdf.url}
\alias{brownianBridgeEmpcdf.url}
\title{brownianBridgeEmpcdf.url}
\format{An object of class \code{character} of length 1.}
\usage{
brownianBridgeEmpcdf.url
}
\description{
Url for downloading the simulated Brownian bridge distribution. 
It is used an empirical quantile function to determine 
p-values in the asymptotic wasserstein test function .wassersteinTestAsy
}
\keyword{datasets}
