% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\name{.brownianBridgeEmpcdf}
\alias{.brownianBridgeEmpcdf}
\title{.brownianBridgeEmpcdf}
\usage{
.brownianBridgeEmpcdf(v)
}
\arguments{
\item{x}{numeric in the Brownian distribution}
}
\value{
Value at x of the Brownian distribution
}
\description{
An empirical cumulative distribution function of a simulated Brownian bridge
distribution. It is used an empirical quantile function to determine 
p-values in the asymptotic wasserstein test function wasserstein.test.asy
}
