% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wasserstein_SingleCell.r
\name{.fishersCombinedPval}
\alias{.fishersCombinedPval}
\title{.fishersCombinedPval}
\usage{
.fishersCombinedPval(x)
}
\arguments{
\item{x}{vector of the two p-values that are to be aggregated}
}
\value{
The combined p-value
}
\description{
Aggregates two p-values into a combined p-value according to Fisher’s method
}
\details{
Aggregates two p-values into a combined p-value according to
Fisher’s method.
}
\examples{
x<-c(0.0004,2e-7)
.fishersCombinedPval(x)


}
