% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{.quantileCorrelation}
\alias{.quantileCorrelation}
\title{.quantileCorrelation}
\usage{
.quantileCorrelation(x, y, pr = NULL)
}
\arguments{
\item{x}{numeric vector representing distribution x}

\item{y}{numeric vector representing distribution y}

\item{pr}{probabilities for computing the quantiles of x and y. 
By default is set to 1000 equidistant quantiles starting at q_1 = 0.5/1000.}
}
\value{
quantile-quantile correlation of x and y
}
\description{
Computes the quantile-quantile correlation of x and y, using the quantile
type 1 implementation in R and Pearson correlation.
}
