% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{.relativeError}
\alias{.relativeError}
\title{Compute relative error}
\usage{
.relativeError(x, y)
}
\arguments{
\item{x}{a number}

\item{y}{a number}
}
\value{
The relative error between \eqn{x} and \eqn{y}
}
\description{
Computes the relative error between two numbers \eqn{x} and \eqn{y}.
}
\details{
The relative error \eqn{e} is defined as \eqn{e = | 1 -  x/y |}
}
