% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WassersteinTest.R
\name{.wassPermProcedure}
\alias{.wassPermProcedure}
\title{.wassPermProcedure}
\usage{
.wassPermProcedure(x, y, permnum, seed = NULL)
}
\arguments{
\item{x}{numeric vector representing distribution A}

\item{y}{numeric vector representing distribution B}

\item{permnum}{integer interpreted as the number of permutations to be
performed}

\item{seed}{number to be used as a seed for reproducibility of the
permutation procedure. By default, NULL is given and no seed is set.}
}
\value{
vector with squared 2-Wasserstein distances computed on random
 shuffles of the two input vectors
}
\description{
Permutation Procedure that calculates the squared 2-Wasserstein distance for
random shuffles of two input distributions and returns them as a vector.
}
