% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\docType{data}
\name{empcdf.ref}
\alias{empcdf.ref}
\title{empcdf.ref}
\format{An object of class \code{NULL} of length 0.}
\usage{
empcdf.ref
}
\description{
An empirical cumulative distribution function of a simulated Brownian bridge
distribution. It is used an empirical quantile function to determine 
p-values in the asymptotic wasserstein test function wasserstein.test.asy
}
\keyword{data}
