% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\docType{data}
\name{empcdf.ref}
\alias{empcdf.ref}
\title{An empirical cumulative distribution function based on the values in value.integral}
\description{
An empirical cumulative distribution function based on the values in value.integral
}
\keyword{data}
