% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{permutations}
\alias{permutations}
\title{Permutations
Returns permutations of a given NumericVector as columns in a NumericMatrix
object.}
\usage{
permutations(x, num_permutations)
}
\arguments{
\item{x}{:    NumericVector representing a vector that is to be permutated}

\item{num_permutations}{:     Int representing the number of permutations
that are to be performed.}
}
\value{
a matrix containing in every column one permutations of the input vector
}
\description{
Permutations
Returns permutations of a given NumericVector as columns in a NumericMatrix
object.
}
