% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{permutations}
\alias{permutations}
\title{Return permutations of a given vector as columns in a matrix}
\usage{
permutations(x, num_permutations)
}
\arguments{
\item{x}{vector that is to be permutated}

\item{num_permutations}{number of permutations to be performed}
}
\value{
a matrix, where each of the \code{num_permutations} columns represents one permutation of the input vector
}
\description{
Returns permutations of a given vector as columns in a matrix
}
\examples{
x <- 1:10
set.seed(24)
permutations(x, 5)

}
