% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{permutations}
\alias{permutations}
\title{permutations}
\usage{
permutations(x, num_permutations)
}
\arguments{
\item{x}{NumericVector representing a vector that is to be permutated}

\item{num_permutations}{Integer representing the number of permutations
that are to be performed.}
}
\value{
a matrix containing in every column one permutations of the
	input vector
}
\description{
Returns permutations of a given NumericVector as columns in a NumericMatrix
object.
}
\examples{
x <- seq(1:10)
m <- permutations(x, 5)
dim(m)
#[1] 10  5

}
