% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{squared_wass_approx}
\alias{squared_wass_approx}
\title{squared_wass_approx}
\usage{
squared_wass_approx(x, y, p = 1)
}
\arguments{
\item{x}{Vector representing an empirical distribution under condition A}

\item{y}{Vector representing an empirical distribution under condition B}

\item{p}{exponent of the wasserstine distance}
}
\value{
The approximated squared wasserstein distance between x and y
}
\description{
Approximation of the squared wasserstein distance.
Calculation based on the mean squared difference between the equidistant 
quantiles of the two input vectors a and b.
As an approximation of the distribution, 1000 quantiles are computed for each vector.
}
