% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{squared_wass_decomp}
\alias{squared_wass_decomp}
\title{squared_wass_decomp}
\usage{
squared_wass_decomp(x, y, p = 1)
}
\arguments{
\item{x}{Vector representing an empirical distribution under condition A}

\item{y}{Vector representing an empirical distribution under condition B}

\item{p}{exponent of the wasserstine distance}
}
\value{
An named Rcpp::List with the wasserstein distance between x and y,
   decomposed into terms for size, location, and shape
}
\description{
Approximation of the squared Wasserstein distance between two vectors,
decomposed into size, location and shape. 
Calculation based on the mean squared difference between the equidistant 
quantiles of the two input vectors a and b.
As an approximation of the distribution, 1000 quantiles are computed for each vector.
}
