% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WassersteinSingleCell.R
\docType{methods}
\name{wasserstein.sc}
\alias{wasserstein.sc}
\alias{wasserstein.sc,matrix,vector-method}
\alias{wasserstein.sc-method,matrix,vector,ANY,ANY,ANY-method}
\alias{wasserstein.sc,SingleCellExperiment,SingleCellExperiment-method}
\alias{wasserstein.sc,SingleCellExperiment,SingleCellExperiment,ANY,ANY,ANY-method}
\title{wasserstein.sc}
\usage{
wasserstein.sc(x, y, method = c("TS", "OS"), permnum = 10000,
  seed = NULL)

\S4method{wasserstein.sc}{matrix,vector}(x, y, method = c("TS", "OS"),
  permnum = 10000, seed = NULL)

\S4method{wasserstein.sc}{SingleCellExperiment,SingleCellExperiment}(x, y,
  method = c("TS", "OS"), permnum = 10000, seed = NULL)
}
\arguments{
\item{x}{matrix of single-cell RNA-sequencing expression data with genes in
rows and samples (cells) in columns}

\item{y}{vector of condition labels}

\item{method}{method employed in the testing procedure: “OS” for the
one-stage method (i.e. semi-parametric testing applied to all (zero and
non-zero) expression values); “TS” for the two-stage method (i.e.
semi-parametric testing applied to non-zero expression values only, combined
with a separate testing for differential proportions of zero expression
using logistic regression). If this argument is not given, a two-sided test
is run by default.}

\item{permnum}{number of permutations used in the permutation testing
procedure. If this argument is not given, 10000 is used as default}

\item{seed}{number to be used as a seed for reproducibility of the
permutation procedure. By default, NULL is given and no seed is set.}
}
\value{
See the corresponding values in the description of the function
.testWass, where the argument inclZero=TRUE in .testWass has to be
identified with the argument method=”OS”, and the argument inclZero=FALSE in
.testWass with the argument method=”TS”.
A vector concerning the testing results, precisely (see Schefzik and
Goncalves (2019) for details) in case of inclZero=TRUE:
\itemize{
\item d.wass: 2-Wasserstein distance between the two samples computed
 by quantile approximation
\item d.wass^2: squared 2-Wasserstein distance between the two samples
 computed by quantile approximation
\item d.comp^2: squared 2-Wasserstein distance between the two samples
 computed by decomposition approximation
\item d.comp: 2-Wasserstein distance between the two samples computed by
 decomposition approximation
\item location: location term in the decomposition of the squared
 2-Wasserstein distance between the two samples
\item size: size term in the decomposition of the squared 2-Wasserstein
 distance between the two samples
\item shape: shape term in the decomposition of the squared 2-Wasserstein
 distance between the two samples
\item rho: correlation coefficient in the quantile-quantile plot
\item pval: p-value of the semi-parametric 2-Wasserstein distance-based
 test
\item p.ad.gpd in case the GPD fitting is performed: p-value of the
 Anderson-Darling test to check whether the GPD actually fits the data well
 (otherwise NA)
\item N.exc: in case the GPD fitting is performed: number of exceedances
 (starting with 250 and iteratively decreased by 10 if necessary) that are
 required to obtain a good GPD fit (i.e. p-value of Anderson-Darling test
 greater or eqaul to 0.05)(otherwise NA)
\item perc.loc: fraction (in %) of the location part with respect to the
 overall squared 2-Wasserstein distance obtained by the decomposition
 approximation
\item perc.size: fraction (in %) of the size part with respect to the
 overall squared 2-Wasserstein distance obtained by the decomposition
 approximation
\item perc.shape: fraction (in %) of the shape part with respect to the
 overall squared 2-Wasserstein distance obtained by the decomposition
 approximation
\item decomp.error: relative error between the squared 2-Wasserstein
 distance computed by the quantile approximation and the squared
 2-Wasserstein distance computed by the decomposition approximation
\item pval.adj: adjusted p-value of the semi-parametric 2-Wasserstein
 distance-based test according to the method of Benjamini-Hochberg
}
In case of inclZero=FALSE:
\itemize{
\item d.wass: 2-Wasserstein distance between the two samples computed
 by quantile approximation
\item d.wass^2: squared 2-Wasserstein distance between the two samples
 computed by quantile approximation
\item d.comp^2: squared 2-Wasserstein distance between the two samples
 computed by decomposition approximation
\item d.comp: 2-Wasserstein distance between the two samples computed by
 decomposition approximation
\item location: location term in the decomposition of the squared
 2-Wasserstein distance between the two samples
\item size: size term in the decomposition of the squared 2-Wasserstein
 distance between the two samples
\item shape: shape term in the decomposition of the squared 2-Wasserstein
 distance between the two samples
\item rho: correlation coefficient in the quantile-quantile plot
\item p.nonzero: p-value of the semi-parametric 2-Wasserstein distance-based
 test (based on non-zero expression only)
\item p.ad.gpd: in case the GPD fitting is performed: p-value of the
 Anderson-Darling test to check whether the GPD actually fits the data well
 (otherwise NA)
\item N.exc: in case the GPD fitting is performed: number of exceedances
 (starting with 250 and iteratively decreased by 10 if necessary) that are
 required to obtain a good GPD fit (i.e. p-value of Anderson-Darling test
 greater or eqaul to 0.05)(otherwise NA)
\item perc.loc: fraction (in %) of the location part with respect to the
 overall squared 2-Wasserstein distance obtained by the decomposition
 approximation
\item perc.size: fraction (in %) of the size part with respect to the
 overall squared 2-Wasserstein distance obtained by the decomposition
 approximation
\item perc.shape: fraction (in %) of the shape part with respect to the
 overall squared 2-Wasserstein distance obtained by the decomposition
 approximation
\item decomp.error: relative error between the squared 2-Wasserstein
 distance computed by the quantile approximation and the squared 
 2-Wasserstein distance computed by the decomposition approximation
\item p.zero: p-value of the test for differential proportions of zero
 expression (logistic regression model)
\item p.combined: combined p-value of p.nonzero and p.zero obtained by
 Fisher’s method
\item p.adj.nonzero: adjusted p-value of the semi-parametric 2-Wasserstein
 distance-based test (based on non-zero expression only) according to the
 method of Benjamini-Hochberg
\item p.adj.zero: adjusted p-value of the test for differential proportions
 of zero expression (logistic regression model) according to the method of
 Benjamini-Hochberg
\item p.adj.combined: adjusted combined p-value of p.nonzero and p.zero
 obtained by Fisher’s method according to the method of Benjamini-Hochberg
}
}
\description{
Two-sample test for single-cell RNA-sequencing data to check for differences
between two distributions (conditions) using the 2-Wasserstein distance:
Semi-parametric implementation using a permutation test with a generalized
Pareto distribution (GPD) approximation to estimate small p-values
accurately
}
\details{
Details concerning the permutation testing procedures for
single-cell RNA-sequencing data can be found in Schefzik and Goncalves
(2019). Corresponds to the function .testWass when identifying the argument
inclZero=TRUE in .testWass with the argument method=”OS” and the argument
inclZero=FALSE in .testWass with the argument method=”TS”.
}
\examples{

# some data in two conditions
cond1 <- matrix(rnorm(100, 42, 1), nrow=1)
cond2 <- matrix(rnorm(100, 45, 3), nrow=1)

# call wasserstein.sc with a matrix
# and a vector denoting conditions
dat <- cbind(cond1, cond2)
condition <- c(rep(1, 100), rep(2, 100))
wasserstein.sc(dat, condition, "TS", 100)

# call wasserstein.sc with two SingleCellExperiment objects
sce1 <- SingleCellExperiment::SingleCellExperiment(
           assays=list(counts=cond1, logcounts=log10(cond1)))
sce2 <- SingleCellExperiment::SingleCellExperiment(
           assays=list(counts=cond2, logcounts=log10(cond2)))
wasserstein.sc(sce1, sce2, "TS", 100)


}
\references{
Schefzik and Goncalves (2019).
}
