% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WassersteinSingleCell.R
\name{wasserstein.sc}
\alias{wasserstein.sc}
\title{wasserstein.sc}
\usage{
wasserstein.sc(x, y, ...)
}
\arguments{
\item{x}{matrix of single-cell RNA-sequencing expression data with genes in
rows and samples (cells) in columns}

\item{y}{vector of condition labels}

\item{permnum}{number of permutations used in the permutation testing
procedure}

\item{method}{method employed in the testing procedure: “OS” for the
one-stage method (i.e. semi-parametric testing applied to all (zero and
non-zero) expression values); “TS” for the two-stage method (i.e.
semi-parametric testing applied to non-zero expression values only, combined
with a separate testing for differential proportions of zero expression
using logistic regression)}
}
\value{
See the corresponding values in the description of the function
.testWass, where the argument inclZero=TRUE in .testWass has to be
identified with the argument method=”OS”, and the argument inclZero=FALSE in
.testWass with the argument method=”TS”.
}
\description{
Two-sample test for single-cell RNA-sequencing data to check for differences
between two distributions (conditions) using the 2-Wasserstein distance:
Semi-parametric implementation using a permutation test with a generalized
Pareto distribution (GPD) approximation to estimate small p-values
accurately
}
\details{
Details concerning the permutation testing procedures for
single-cell RNA-sequencing data can be found in Schefzik and Goncalves
(2019). Corresponds to the function .testWass when identifying the argument
inclZero=TRUE in .testWass with the argument method=”OS” and the argument
inclZero=FALSE in .testWass with the argument method=”TS”.
}
\examples{

# some data in two conditions
cond1 <- matrix(rnorm(100, 42, 1), nrow=1)
cond2 <- matrix(rnorm(100, 45, 3), nrow=1)

# call wasserstein.sc with a matrix
# and a vector denoting conditions
dat <- cbind(cond1, cond2)
condition <- c(rep(1, 100), rep(2, 100))
wasserstein.sc(dat, condition, 100, method="TS")

# call wasserstein.sc with two SingleCellExperiment objects
sce1 <- SingleCellExperiment(
           assays=list(counts=cond1, logcounts=log10(cond1)))
sce2 <- SingleCellExperiment(
           assays=list(counts=cond2, logcounts=log10(cond2)))
wasserstein.sc(sce1, sce2, 100, method="TS")

}
\references{
Schefzik and Goncalves (2019).
}
