% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wasserstein_metric}
\alias{wasserstein_metric}
\title{wasserstein_metric}
\usage{
wasserstein_metric(a_, b_, p = 1, wa_ = NULL, wb_ = NULL)
}
\arguments{
\item{a_}{NumericVector representing an empirical distribution under condition A}

\item{b_}{NumericVector representing an empirical distribution under condition B}

\item{p}{exponent of the wasserstine distance}

\item{wa_}{NumericVector representing the weights of datapoints (interpreted as clusters) in a}

\item{wb_}{NumericVector representing the weights of datapoints (interpreted as clusters) in b}
}
\value{
The wasserstein (transport) distance between a and b
}
\description{
wasserstein_metric
}
