% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wasserstein_metric}
\alias{wasserstein_metric}
\title{Wasserstein Metric}
\usage{
wasserstein_metric(a, b, p = 1, wa_ = NULL, wb_ = NULL)
}
\arguments{
\item{a}{: NumericVector a representing a distribution}

\item{b}{: NumericVector b representing a distribution}

\item{p}{: int p representing the exponent in the root mean squared difference}

\item{wa}{: NumericVector wa representing a weight matrix for vector a}

\item{wb}{: NumericVector wb representing a weight matrix for vector b}
}
\value{
a floating point number representing the wasserstein distance or metric
	between the given input distributions
}
\description{
Returns the Wasserstein Metric of two input vectors a and b.
Reimplementation in Cpp of the function wasserstein1d in the package transport.
}
